LPS-105689 Do tag directly without jsp
