+++ modules/apps/change-tracking/change-tracking-web/src/main/java/com/liferay/change/tracking/web/internal/constants/CTWebKeys.java
@@ -28,9 +28,6 @@ public interface CTWebKeys {
 	public static final String CHANGE_LISTS_HISTORY_DISPLAY_CONTEXT =
 		"CHANGE_LISTS_HISTORY_DISPLAY_CONTEXT";
 
-	public static final String CHANGE_TRACKING_INDICATOR_DISPLAY_CONTEXT =
-		"CHANGE_TRACKING_INDICATOR_DISPLAY_CONTEXT";
-
 	public static final String CONFLICT_INFO_MAP = "CONFLICT_INFO_MAP";
 
 	public static final String CT_COLLECTION = "CT_COLLECTION";
