+++ modules/apps/change-tracking/change-tracking-web/src/main/resources/META-INF/resources/dynamic_include/ChangeTrackingIndicator.js
