--- modules/apps/change-tracking/change-tracking-web/src/main/resources/META-INF/resources/dynamic_include/change_tracking_indicator.jsp
@@ -1,24 +1,0 @@
-<%--
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
---%>
-
-<%@ include file="/dynamic_include/init.jsp" %>
-
-<div class="change-tracking-indicator">
-	<react:component
-		data="<%= changeTrackingIndicatorDisplayContext.getReactData() %>"
-		module="dynamic_include/ChangeTrackingIndictator"
-	/>
-</div>
